+++ doc/source/doxygen-docs/changelog.md
@@ -39,6 +39,9 @@
       - New method to render into a depth image mrpt::opengl::CFBORender::render_RGBD().
     - mrpt::opengl::CCamera::setProjectiveFromPinhole() now allows defining a camera by means of a pinhole model.
     - New class mrpt::opengl::COpenGLFramebuffer, used to refactor mrpt::opengl::CFBORender
+    - New methods to control face culling:
+      - mrpt::opengl::CRenderizableShaderTriangles::cullFaces()
+      - mrpt::opengl::CRenderizableShaderTexturedTriangles::cullFaces()
   - \ref mrpt_poses_grp
     - New function mrpt::poses::sensor_poses_from_yaml()
     - New header `<mrpt/poses/gtsam_wrappers.h>`, see \ref mrpt_gtsam_wrappers